feat: move page and layout to locale
